fix: added error handling in CRUD routes

+++ src/main/java/com/example/avanadedesafiotecnico/controllers/CharacterController.java
@@ -1,8 +1,8 @@
 package com.example.avanadedesafiotecnico.controllers;
-
 import com.example.avanadedesafiotecnico.entities.Character;
 import com.example.avanadedesafiotecnico.payloads.request.CharacterRequest;
 import com.example.avanadedesafiotecnico.repositories.CharacterRepository;
+import com.example.avanadedesafiotecnico.repositories.ClassRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +14,8 @@ public class CharacterController {
     @Autowired
     private CharacterRepository characterRepository;
 
+    @Autowired
+    private ClassRepository classRepository;
     @GetMapping
     public ResponseEntity findAllCharacters() {
         var allPersonagem = characterRepository.findAll();
@@ -24,24 +26,39 @@ public class CharacterController {
     public ResponseEntity findCharacterById(@PathVariable Long id) {
         var character = characterRepository.findById(id);
         if (character.isEmpty()) {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Esse Id nao pertence a nenhum personagem!");
         }
         return ResponseEntity.ok(character.get());
     }
 
     @PostMapping
-    public ResponseEntity<Character> createCharacter(@RequestBody CharacterRequest characterRequestDTO) {
-        var character = new Character();
-        character.setName(characterRequestDTO.getName());
-        character.setClass_id(characterRequestDTO.getClass_id());
+    public ResponseEntity createCharacter(@RequestBody CharacterRequest characterRequestDTO) {
+       var classExists = classRepository.findById(characterRequestDTO.getClass_id());
+        System.out.println(classExists);
 
-        characterRepository.save(character);
+       if (classExists.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Esse Id nao pertence a nenhuma classe existente!");
+       }
 
-        return ResponseEntity.status(HttpStatus.CREATED).body(character);
+       if (characterRequestDTO.getName().isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("O nome do personagem nao pode ser vazio!");
+       }
+
+       var character = new Character();
+       character.setName(characterRequestDTO.getName());
+       character.setClass_id(characterRequestDTO.getClass_id());
+
+       characterRepository.save(character);
+
+       return ResponseEntity.status(HttpStatus.CREATED).body(character);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity deleteCharacter(@PathVariable Long id) {
+        var character = characterRepository.findById(id);
+        if (character.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Esse Id nao pertence a nenhum personagem!");
+        }
         characterRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
@@ -50,13 +67,13 @@ public class CharacterController {
     public ResponseEntity updateCharacters(@PathVariable Long id, @RequestBody CharacterRequest characterRequestDTO) {
         var character = characterRepository.findById(id);
         if (character.isEmpty()) {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Esse Id nao pertence a nenhum personagem!");
         }
+
         character.get().setName(characterRequestDTO.getName());
         character.get().setClass_id(characterRequestDTO.getClass_id());
 
         characterRepository.save(character.get());
-
-        return ResponseEntity.noContent().build().ok(character.get());
+        return ResponseEntity.status(HttpStatus.CREATED).body(character);
     }
 }
